+++ px7_music/config.py
@@ -36,5 +36,6 @@ YTDLP_STREAM_OPTS = {
 
 from pathlib import Path
 
-FAV_FILE = Path.home() / ".px7_favorites.json"
-PL_FILE = Path.home() / ".px7_playlists.json"
+LEGACY_FAV_FILE = Path.home() / ".px7_favorites.json"
+FAV_FILE = Path.home() / ".px7" / ".px7_favorites.json"
+PL_FILE = Path.home() / ".px7" / ".px7_playlists.json"